refactor: Make the sort fcs function not a hook
Some reasons:
- The useCallback and useMemo in the hook had no effect as the input
callback function was recreated on every render. Also the function
needs the now variable, which is updated every second, so the
memoisation would be limited even if it was working.
- Making it a utility function removed the need for lots of mocks in
the test class.
- The sorting is pretty quick. On simulator (which is pretty slow) it
was measured to be less than 0.1ms per fare contract.
