+++ src/fare-contracts/__tests__/sort-fc-or-reservation-by-validity-and-creation.test.ts
@@ -1,42 +1,8 @@
-import {
-  ValidityStatus,
-  useSortFcOrReservationByValidityAndCreation,
-} from '../utils';
+import type {ValidityStatus} from '../utils';
 import {FareContract, Reservation, TravelRight} from '@atb/ticketing/types';
 
-import {LoadingParams} from '@atb/loading-screen/types';
-import React from 'react';
 import {addMinutes} from 'date-fns';
-
-const DEFAULT_MOCK_STATE: LoadingParams = {
-  isLoadingAppState: false,
-  authStatus: 'authenticated',
-  firestoreConfigStatus: 'success',
-  remoteConfigIsLoaded: true,
-};
-
-jest.mock('@atb/auth/AuthContext', () => {});
-jest.mock('@atb/mobile-token', () => {});
-jest.mock('@atb/ticketing/TicketingContext', () => {});
-jest.mock('@atb/configuration/FirestoreConfigurationContext', () => {});
-jest.mock('@atb/api', () => {});
-jest.mock('@atb/time', () => {});
-jest.mock('@react-native-firebase/remote-config', () => {});
-jest.mock('@entur-private/abt-mobile-client-sdk', () => {});
-jest.mock('@bugsnag/react-native', () => {});
-jest.mock('@react-native-firebase/auth', () => {});
-jest.mock('@entur-private/abt-token-server-javascript-interface', () => {});
-jest.mock('react-native-device-info', () => {});
-jest.mock('react-native-inappbrowser-reborn', () => {});
-jest.mock('@atb/auth', () => ({
-  useAuthContext: () => ({
-    authStatus: DEFAULT_MOCK_STATE,
-    abtCustomerId: '1',
-    retryAuth: () => {},
-  }),
-}));
-jest.spyOn(React, 'useCallback').mockImplementation((f) => f);
-jest.spyOn(React, 'useMemo').mockImplementation((f) => f());
+import {sortFcOrReservationByValidityAndCreation} from '../sort-fc-or-reservation-by-validity-and-creation';
 
 type MockedFareContract = FareContract & {
   validityStatus: ValidityStatus;
@@ -94,7 +60,8 @@ describe('Sort by Validity', () => {
       mockupReservation('3', 'valid', 0),
     ];
 
-    const result = useSortFcOrReservationByValidityAndCreation(
+    const result = sortFcOrReservationByValidityAndCreation(
+      '',
       now,
       fcOrReservations,
       (_, fareContract) => (fareContract as MockedFareContract).validityStatus,
@@ -118,7 +85,8 @@ describe('Sort by Validity', () => {
       mockupFareContract('2', 'valid', 0),
     ];
 
-    const result = useSortFcOrReservationByValidityAndCreation(
+    const result = sortFcOrReservationByValidityAndCreation(
+      '',
       now,
       fcOrReservations,
       (_, fareContract) => (fareContract as MockedFareContract).validityStatus,
@@ -143,7 +111,8 @@ describe('Sort by Validity', () => {
       mockupFareContract('4', 'valid', 0.11),
     ];
 
-    const result = useSortFcOrReservationByValidityAndCreation(
+    const result = sortFcOrReservationByValidityAndCreation(
+      '',
       now,
       fcOrReservations,
       (_, fareContract) => (fareContract as MockedFareContract).validityStatus,
